// frontend/src/backend.d.ts
export interface Treatment {
    date: Time;
    plan?: TreatmentPlan;
    type: string;
    notes: string;
    price: bigint;
}
export interface Location {
    latitude: number;
    longitude: number;
    timestamp: Time;
}
export type Time = bigint;
export interface Spending {
    id: string;
    date: Time;
    description: string;
    category: SpendingCategory;
    amount: bigint;
}
export interface PatientSession {
    id: string;
    checkIn: Time;
    duration: bigint;
    visitNumber: bigint;
    progressPercentage: bigint;
    patientId: string;
    date: Time;
    treatmentPackageId: string;
    checkOut?: Time;
}
export interface MiscExpense {
    id: string;
    date: Time;
    description: string;
    category: SpendingCategory;
    amount: bigint;
}
export interface Patient {
    id: string;
    debt: bigint;
    name: string;
    paid: bigint;
    diagnosis: string;
    address: string;
    phone: string;
    treatments: Array<Treatment>;
}
export interface CheckOutData {
    patientId: string;
}
export interface StaffMember {
    id: string;
    name: string;
    qualifications: string;
    phone: string;
    position: string;
    payRate: bigint;
}
export interface TreatmentPlan {
    endDate: Time;
    description: string;
    startDate: Time;
}